requirements.txt
tweepy==3.7.0
tinydb==3.13.0
requests==2.21.0
urllib3==1.24.1
numpy==1.16.2
Pillow==5.4.1

